The CSV import tests use a project by default
The CSV importer tests were using a fund activity as the parent activity of the
activity we're importing by default, therefore creating a programme activity, which
is not how the importer is currently being used. We're therefore correcting this to
better reflect reality and also then allowing us to add an additional test to check
programme activities can be created correctly.