Allow forecasts to be imported into historic reports
This adjusts the forecast import script so that we can import into the
"historic" report for each partner.

"Historic" reports are those that represent all the data we ingested
when onboarding a delivery partner, before they starting reporting
through RODA. They do not have a financial quarter. To select these
reports, pass zero for the relevant options:

    $ ruby script/import_forecasts.rb --quarter 0 --year 0 ...