bin/db-restore fixes
The command wasn't working correctly because it was attempting to seed
the `users` table that didn't exist as it was explicitly excluded from
the database dump.

Changing `--exclude-table` to `--exclude-table-data` means that the
schema for the users table is dumped, but its contents isn't.

I've also excluded the internal `ar_internal_metadata` and
`spatial_ref_sys` tables from the backup as well.

Finally calling `rails db:environment:set` let's Rails know we're
running a development environment (despite the copy of the database
looking like a production one).

NB: I'm entirely sure why this is needed when we're excluding
`ar_internal_metadata` from the dump, but I got warnings without it.