Include headings for forecastable quarters in CSV template
Currently, the CSV template for uploading forecasts only includes
metadata about each activity -- its name and IDs. To submit forecasts,
users have to add column headings for the quarters they wish to report
forecasts for, and these headings must follow the syntax:

    FC <year> FY Q<quarter>

    e.g.:
    FC 2020/21 FY Q4

This is hard to remember and get right and makes the process more
error-prone. So, we are adding column headings for the next 20 quarters
following the current report, to make it easier to fill these templates
out.