Allow blank cells in the forecast upload CSV
In the previous commit we made the CSV template include column headings
for twenty upcoming quarters, to avoid users having to type these out
themselves. This raises the question of what value to pre-fill in the
cells under those headings. Our choices are:

1. Set them to 0. This would mean any cells that the user does not
   modify before re-uploading the completed template would write the
   value 0 over whatever value we currently have for that
   activity/quarter.

2. Fill them with the current value we have for that activity/quarter.
   This would mean the user's upload would not overwrite 0 for any cells
   they had not modified, so effectively only the cells they *had*
   modified would cause changes.

3. Leave them blank. This means the user only has to fill in values for
   activities/quarters they actually want to modify, but requires the
   service to accept blank cells.

The key problem with options 1 and 2 is they give the importer no way to
differentiate between data the user has deliberately set, and cells
they've left unmodified. Making 0 the default and therefore overwriting
existing data seems "obviously" bad so we can discard option 1.

Option 2 is more subtle. It works as long as only one person is
performing the task of downloading the template, making changes, and
re-uploading, at the same time. If two users both download the template,
make different changes to it, and then upload their CSVs, they will
overwrite each other's changes, because the importer treats *all* cells
in the uploaded file as current data.

Option 3 means we can distinguish which cells the user has actually
modified, so they can be selective about which forecasts they are
actively asserting, without having to have or fill out complete up to
date forecasts for all other activities and quarters they are not
currently interested in. It eliminates the problems with options 1 and
2, at the cost that an empty cell is no longer considered an error. We
think the risk of this allowing mistakes to happen is relatively low,
and certainly won't cause damage since blank cells are ignored and don't
lead to any existing data being overwritten or deleted.

Therefore we leave all the forecast cells blank, and allow the service
to accept empty cells as input without complaining, only saving the
forecasts that are actually present.