Level and parent are always required
The `level` and `parent` attributes will always be set now and their
form steps have been removed, so we can remove the validation contexts which
means they are always validated.

This lead to `shoulda_matchers` complaining when testing
`belongs_to(:parent).optional`, presumably because of the
`unless: proc { ...}` statement. I couldn't figure out a way to make
this work, so I opted for adding tests of the parent validations at
every level instead.

Because `level` is now a required attribute, we can enforce this on
the database too

As a nice side effect of doing this, Pundit no longer needs to apply
special checks to allow users to edit an activity without a level. This
was always a bit weird, because `level` is heavily used when checking
permissions.