Add total spend to activity financials page
To help users reconcile the amount spent on an activity against a given
budget. We don't flag any variance because RODA is only a reporting
service, it is up to user to solve any errors.

Use the ActivityPresenter on the financials page so that we can access
the appropriate helper methods.