Only set provided variables when updating
When we're updating an activity, we don't always want to have to set every single
variable. As it stands, when we update a record, there are some variables we're
expected to set every single time (e.g. sector code).

This updates the Converter to behave differently when updating to when creating a
new activity. If we're updating, we grab the row, and only run the converter on
fields that are expressly filled out. We also don't do any of the inferrence when
updating, as this can overwrite data that is already set on the Activity.

This also has the side effect of any policy marker fields not being set to
`not_assessed` by default. This is fine for now, and we're going to write a seperate
card to write a migration that sets all Activity policy marker fields to default to
`not_assessed` at the database level.