Collect more budget types
In order to support different funding arrangements we need to record
different types of budget.

The two simplest types are transferred and external income, external
income has two distinctions, budget from an source that is considered
official development assistance and from a source that is not.

So that users can let us know what type of budget they are telling us
about, we collect them on budgets.

The existing `funding_type` code list, is really where we derive the two
'funds' (GCRF and Newton), whilst this feels similar it is not the same
thing. Therefore we introduce a new BEIS budget_type code list to store
the BEIS defined budget types (as opposed to the IATI ones which means
something else).

We load the code lists as constants on instances of Budget so we can
validate the budget types, it also lets us translate from the coded
value to a human readable name and back again without having to
duplicate the values in the locales files.