Refactor handling of implementing organisations
In certain circumstances, when trying to update an activity, we hit an error in the
ImplemeningOrganisationBuilder, where the builder tried to add an error to the
Errors array, but because there were no implementing organisations available, we
got an `undefined method `errors' for nil:NilClass` error.

This reworks how the `ImplemeningOrganisationBuilder` works to memoize the
organisation that gets built, rather than fetching the organisation's last
implementing organisation, which may or may not be the correct one, or even
exist at all.

I've also updated the updater, so it only attempts to build an implementing
organisation if the relevant fields are present, in a similar way to how the creator
works.