Remove policy markers from ALLOWED_BLANK_FIELDS
Now the default value of the policy markers is assumed to be `not_assessed`, we
can remove these values from the `ALLOWED_BLANK_FIELDS` array in the
Activity importer. I've also updated the tests, as now when a policy marker field is
blank, it's assumed to have not been assessed.