Use budget_type not funding_type
Previous work de-coupled the IATI concept of budget_type from the
database column `budget_type`. We can now use this column to store the
actual budget type as opposed to the less well named 'funding type'.

At this time, all budget in the application is considered 'direct' from
either GCRF or Newton fund so we can copy the parent activity's
`source_fund_code` and rely on the fact that the coded value is the same
for Newton and GCRF funds (1 and 2) and Newton and GCRF budget (1 and 2).

We loose the existing record of the IATI budget type, but we have
already made this always return 'committed', 2 and BEIS have made the
decision that this is the correct IATI budget type for their use case.