Generate sector category hint with helper
Generate the guidance link with the new helper.

The `level` parameter was removed from the sector category hint in commit 708909b1ed0223676403d0665bcff63e052fbebd, so it can be safely removed from the view.

Add the magic `_html` suffix to the translation key, to remove the need to call `html_safe` on it.

Co-authored-by: robbiepaul <robbie.paul@dxw.com>