Correct test for empty forecast cells in CSV template
PR #1032 corrected for the fact that this test assumes the acitivty rows
in the CSV template come out in a defined order, which is not true. What
this test should actually be doing is checking that every activity in
the CSV has a row full of empty cells for the forecast columns.