Remove unnecessary specs in ActivityPolicy
As we now always check that there is an editable report based on both
the user's organisation and the activity's fund, we can remove the test
cases where there are reports that match the organisation but not the
fund.

I've also changed it so a report *always* exists, with only it's state
changing from 'approved' (not editable) to 'active' (editable) where
appropriate.