More views use ActivityPolicy#create_child?
Update Organisation#show and the third-party projects shared partial to
use the `ActivityPolicy#create_child?` check.

In order to make this work, I had to update the feature specs to set the
extending_organisation to reflect how projects are initialised in
`ActivityDefaults`. This was needed, as `create_child?` uses this
attribute to determine that the current user can create a child activity
under the current one.

Additionally, I updated the report existence check in `ActivityPolicy`
to use the id columns rather than the associations (e.g.
`organisation_id` instead of `organisation`). Without this, you get
Unoptimised Query warnings from bullet, which aren't helpful because you
shouldn't have to remember to include the `extending_organisation`
association whenever you'll want to perform a policy check.