Document the process for importing new data
When we on-board new delivery partners we need to bring in all their
existing activities and financial data at the time they are on-boarded.

This is a one time only operation run by developers with data supplied
by BEIS.

We want to document the process in the code as the most useful place to
do so and expect it to be removed once the service completes on-boarding
all delivery partners in Q1 2021-2022.