Use codelists to initialize policy marker enums
As we want the codelists the be the canonical source of the data, rather than
anywhere else in the code, this changes the `POLICY_MARKER_CODES` to be
sourced from the codelist and initialized when the application boots. This requires
an additional `hash_of_integer_coded_names` method in the Codelist model, so
the codes we return are actual integers, not numbers as strings.

Because the desertification policy markers are slightly different from the main
policy markers, we have a different constant for those.

I have also had to change one of the desertification options to
`principal_objective_and_in_support_of_an_action_programme` from the shorter
`principal_objective_and_in_support`, but because the underlying integer that is
stored in the database hasn't changed, this won't require a data migration.