Add script for converting import data files
To perform the initial import of new delivery partners' records into
RODA, we need to take the spreadsheets they've provided and convert them
into the schema supported by the transaction and forecast importers.

To run the script:

    $ ruby script/convert_import.rb \
        --level C \
        --input path/to/input_file.csv \
        --output path/to/output

This reads the data from `path/to/input_file.csv` and writes two new
files: `path/to/output_transactions.csv` and
`path/to/output_forecasts.csv`, which will be compatible with our
existing importers. It does not write anything to the database or
interact with the application in any other way, it purely converts one
CSV file into two others.

The `--level` option must be either `C` or `D` to indicate the level of
the activities being imported.

The main job of this converter is to recognise the various ways that
identifiers, transactions and forecasts are represented in the files
we've been provided, and canonicalise them.

Identifiers are presented as two columns: `Parent RODA ID` and `RODA ID
Fragment`, as is usual for creating new activities. To build the RODA ID
of the new activity we combine these fields with either `-` at level C
or the empty string at level D.

Transactions are presented column-wise, whereas our importer expects
them row-wise, with these standard columns:

    Activity RODA Identifier
    Financial Year
    Financial Quarter
    Value

So we convert a single input row into multiple possible output rows by
matching column headings that represent transactions. We recognise the
following formats that we have observed in files provided to us:

    Act <yyyy>/<yy> FY Q<q> (<month>, <month>, <month>)
    Act <yyyy>/<yy> Q<q>
    Q<q> <yyyy>-<yyyy> actuals

We extract the year and quarter from these and emit one row of output
for each column we find in the input.

Forecasts are presented column-wise and are also consumed column-wise by
our importer, but there are many different heading formats in the input
we've been given. They include:

    FC Q<q> <yyyy>-<yy>
    FC Q<q> <yyyy>
    FC <yyyy>/<yy> FY Q<q> (<month>, <month>, <month>)
    Q<q> <yyyy>-<yyyy> forecast

We recognise all these and emit a CSV with headings in the form
recognised by our importer:

    FC <yyyy>/<yy> FY Q<q>

Since this set of output columns is dynamic and we're emitting a single
CSV, we need to make sure all rows are emitted with the same set of
quarters in the same order. To achieve this, we use the first row of the
input to extract a set of `forecast_mappings` -- a list of the columns
we plan to emit along with their source column names in the input. We
pass these mappings in to convert each data row.