Change GCRF stategic area checkboxes to integers
This was  also causing issues where the `gcrf_strategic_area_options` were
returning the code as an integer, but the codes were stored as a string. I looked
into tweaking how the options were presented, or even changing the underlying
YAML to made the codes into strings, but I finished up taking the approach that we
should trust the codelist and if the codes are integers, so the data should be
stored as an integer too.