Remove byte order mark from the start of import files
Our script for extracting transactions and forecasts from import data
files was failing to recognise the first heading in CSV files because of
the presence of a byte order mark. We remove this before parsing.

This does mean we need to read the file into memory rather than letting
`CSV.open` read the file however it likes, but the files in question
aren't big enough to consider this a problem.