Load the application environment in `script/convert_import.rb`
The ImportConverter service uses some Rails extensions like
`Object#blank?` and so, even though it doesn't call on any of our
application logic, it's simplest to load the application so that the
code runs in a Rails environment.

We decided to do this, rather than remove Rails-isms from
ImportConverter, because it's easier to work on the code if it runs in
an environment just like everything else, rather than having to remember
it runs in a non-Rails environment and so shouldn't use Rails
extensions.

This is especially true given that all tests run with the application
loaded, and so cannot catch issues with code that needs to run outside a
Rails environment.