Raise an error if identifier columns are not present in import CSVs
We had a problem with the import converter script not recognising the
`RODA ID Fragment` column in an import CSV. It turns out this is because
the file had a byte order mark at the start, which ended up being part
of the first column's header, so the header was actually `\u{FEFF}RODA
ID Fragment`.

Using `values_at` here means we get a `nil` return value for missing
headers rather than throwing an error. To better identify problems with
missing headers, we'd rather throw an error if these required headers
are missing, and using `fetch` gives us that behaviour.