Hide create child button when there is no report
If a user clicks on the button and there is no editable report, the user will see a
"not authorised" message, so this just changes the frontend to better reflect
reality.

I've also had to preload the parent and the subparent in the controller, as the
`editable_report?` check in the Activity policy needs to check the fund.