Make sure the destination field has a value
Before bumping the version of `govuk_design_system_formbuilder` to 2.5.0, this
passed, but we were still passing in a blank value. We now make sure our "fake"
activity has an actual RODA identifier, and check the field is populated with the
value the user entered when presenting the error back to the user.