Refactor report list generation in ReportsController
While working on a new report CSV download, I noticed a lot of
duplication in the methods in this controller, with two copies of each
method that differ only on which associations are eager-loaded. I
refactored to define a single version of each method that accepts an
optional argument of additional associations.

This is intended to make it easier to reuse these methods for cases
where we need to generate new output formats off the same set of reports
for BEIS users.