Generate CSV headers without an activity in the spending breakdown
To generate the header row of a CSV download, we need to be able to run
ActivitySpendingBreakdown with no Activity, because the headers need to
be provided separately from the data, and must be generated even if
there are no activities to report on.

Here we restructure ActivitySpendingBreakdown to make it possible to
generate just the headers without the corresponding values, without
disrupting the `combined_hash` method that makes it easier to test
specific pieces of the data row this class generates.