Record providing org fields for external budgets
The providing organisation name and type are required for external budgets.

The budget type is considered the source of truth that informs
which type of providing organisation we can record. We clear out
any fields that are incompatible with the budget type.

In future commits we plan to enhance the form with scripting that
will hide any irrelevant fields respective to the budget type, but
this ensures that we don't record contradictory information even
if the irrelevant fields are accessible.

Co-authored-by: robbiepaul <robbie.paul@dxw.com>