Search for activities by RODA identifier
This begins a service that supports searching for activities, with the
aim of implementing a simple search box in the UI. It would take a
single string as input and search for activities by matching this string
against various columns.

We start by implementing just a literal match against the RODA
identifier, either the complete identifier or the activity's own
fragment.

For BEIS users, this can query all activities. For delivery partners,
search is restricted to only those activities owned by their
organisation.