Refactor activity search query
The structure of this query is very repetitive and worth refactoring
before we refine it further. We're essentially matching the query string
against several different columns and combining all these matches with
OR.

This separates the code into the list of columns we want to search on,
and then logic for constructing the conditions for each column and
combining them.