Make sure activity fields have all translations
This adds a shared example that covers all presenter fields that have a translation.
It runs through all the codes in a given codelist and checks that the output does
not match the `/translation missing/` regex. I did consider raising an error, but I
was concerned that might cause issues with legacy data. This at least makes sure
any data that we add with the codelists we have now (and any codelists that
change going forward) are covered by the translations.