Support case-insensitive searches
Identifiers in RODA in general are mixed-case and users may be searching
based on a name they remember or which has been relayed to them
verbally, rather than from text they can see on screen. If we expand the
search to other fields like title and description, it's even more
important to ignore case in the input, instead of expecting users to
remember exactly how something is cased in our database.

Here we change the existing query to support case-insensitive matching
by using the `LOWER()` function across all fields we search on.