Only total direct and transferred budgets
This filters the budgets by direct and transferred budget type. I also got a bit
mixed up and concerned that we were just checking magic numbers in the tests,
so rather than creating the budgets and basing our tests on what we think the
totals should be, I'm storing the budgets as variables (with random values) and
totalling up the values based on where they should be in the hierarchy, which gives
us more confidence we're getting the correct budgets for an activity, and the totals
aren't wrong by accident.