Add TableTreeView component for expanding table rows
This allows a rows of a table to be collapsed and expanded based on a
parent to child relationship. When you expand a parent row you see all
the related children rows.

The state of each section is saved to the DOM via the `aria-expanded`
attribute, which also provides accessibility.

This is a modified version of the GOV.UK Accordion component:
https://github.com/alphagov/govuk-frontend/blob/master/src/govuk/components/accordion/accordion.js

Some of the styles are scoped to `.js-enabled` so that the tree view
table will default to showing all rows for non-JavaScript users.

We had to prefix the row's ID attribute with 'activity-' because
ID attributes can't start with numbers and Capybara threw an
InvalidSelector error.