Allow reports without financial year to be accessed
Historical “ingested” reports don’t have any financial year and quarter.

Many of the features built around a report rely on the report having a financial quarter and year, and break if the report doesn’t.

In order to allow the historical reports to be accessed and downloaded, we need to add a lot of guards against the lack of financial year and quarter.

(Any reports being created now enforce the presence of a FY/FQ by assigning default values at creation, and the fields are readonly at database level.)