Remove unsupported `?.` from JavaScript
Despite enabling the `harmony` option in the previous commit, some more
recent JS syntax is still not supported. The optional chaining operator
`?.` is one such feature and `assets:precompile` fails with this error:

    Uglifier::Error: Unexpected token: punc (.)

Ultimately we want to fix this by migrating to Webpacker to compile our
assets, and thereby get access to modern JS features, but this is a
quick fix we can make in the interim to get the site to build.