Install Webpacker
The changes in this commit are the result of adding the `webpacker` [1]
gem, and running `rails webpacker:install`. This gives us the same
config that one would get on a new Rails 6.1 application for compiling
JavaScript.

This doesn't change how existing JavaScript is compiled yet.

We're installing Webpacker so that we can use modern JavaScript syntax
that's not supported by Uglify, or by some browsers we need to support,
such as Internet Explorer.

[1]: https://github.com/rails/webpacker