Show implementing organisations in the report csv
Users want to see the implementing organisations in the report csv. As
there can be multiple implementing organisations for a given activity we
need a separator when outputting into a single csv cell. As with the
rest of the application we use a pipe "|" symbol to achieve this.

As a general pattern we use a presenter to change the way data is
presented in certain contexts, here the csv presenter does the job.