Remove the reporting_organisation association
This removes the `reporting_organisation` `belongs_to` association from the
Activity model and instead sets the first part of the RODA identifier to be
the value of the `SERVICE_OWNER_IATI_REFERENCE`, which is hardcoded to be
`GB-GOV-13` (The IATI refefence of BEIS). The only place that
`reporting_organisation` was actually used was in the IATI XML, so I've tweaked
that, so the IATI XML always returns the service owner for the reporting
organisation.

I've also removed an old data migration, and associated spec (which was failing)
which has already been run