Always eager load relationships needed for IATI XML
We only ever call the find_*_activities services with
`eager_load_parent: false` which makes me think we don't need this flag.
Optionally loading relationships has the potential to be a real headache
in our tests. I feel like it's not too expensive to load these in,
especially as they're not `has_many` relationships so they will take a
limited amount of memory. The database queries themselves should still
be fast enough.

Co-authored-by: CristinaRO <cristina@dxw.com>