Add reportable transactions to activities
When reporting programme level activities to IATI we want to include
transactions belonging to all child activities as well. We don't want
an entry for each transaction rather we want to group them by financial
quarter.

Because we want the XML views to respond in the same way as other level
activities we return the summed transactions as new Transaction objects
that aren't persisted. These instances are only intended to be used in
the XML views so I think this is ok.

Co-authored-by: CristinaRO <cristina@dxw.com>