Deleting a forecast is non-destructive
Deleting a forecast is equivalent to setting the latest revision to 0.

Rather than plainly deleting all revisions, different things need to happen depending on whether the forecast is pinned to a report, like in the case of level C and D activities, or not pinned, like in the case of levels A and B.

The forecast history already takes care of all these cases when setting a value to 0, whether through the interface or via bulk upload.

Deletion is a nicer way for users to express the intent to remove that forecast, while the history mechanism remains invisible from the outside.

Co-authored-by: jcoglan <james.coglan@dxw.com>