Make `Activity#total_forecasted` work with PlannedDisbursement versioning
While checking use of the PlannedDisbursement model throughout the
codebase, I noticed the definition of this method sums all the
PlannedDisbursement objects linked to the activity:

    def total_forecasted
      activity_ids = descendants.pluck(:id).append(id)
      PlannedDisbursement.where(parent_activity_id: activity_ids).sum(:value)
    end

This will include all _versions_ of a forecast -- a logical forecast is
identified by the activity and quarter it relates to -- its
`(parent_activity_id, financial_year, financial_quarter)` fields. Each
logical forecast can have multiple versions, each belonging to a
different Report. The implementation above will sum over all versions,
rather than using a single version of each logical forecast.

Instead we want to only sum over the latest values, which is done using
the PlannedDisbursementOverview service:

    def total_forecasted
      activity_ids = descendants.pluck(:id).append(id)
      overview = PlannedDisbursementOverview.new(activity_ids)
      overview.latest_values.sum(:value)
    end

This required changing PlannedDisbursementOverview so that it can accept
multiple Activity IDs, rather than a single Activity. This makes things
more complicated because versioning works differently for level A/B
activities vs C/D ones. Before this change, this is the query used to
fetch the latest versions of forecasts for a single C/D activity.

    SELECT DISTINCT ON (
      planned_disbursements.financial_year,
      planned_disbursements.financial_quarter
    )
    planned_disbursements.*
    FROM planned_disbursements
    INNER JOIN reports ON reports.id = planned_disbursements.report_id
    WHERE
      planned_disbursements.parent_activity_id = ?
    ORDER BY
      planned_disbursements.financial_year ASC,
      planned_disbursements.financial_year ASC,
      reports.financial_year DESC NULLS LAST,
      reports.financial_quarter DESC NULLS LAST,
      reports.created_at DESC

Let's break down how this works. We select the `planned_disbursements`
records for a given `parent_activity_id`:

    SELECT ...
    FROM planned_disbursements
    WHERE
      planned_disbursements.parent_activity_id = ?

We filter this down to a single record for each financial quarter --
this is what picks a single version from all the available forecasts for
an activity and quarter.

    SELECT DISTINCT ON (
      planned_disbursements.financial_year,
      planned_disbursements.financial_quarter
    )
    ...
    ORDER BY
      planned_disbursements.financial_year ASC,
      planned_disbursements.financial_year ASC,
      ...

And we choose which version to pick by ordering on the financial quarter
of the associated Report, with the most recent one first and historic
reports (no financial quarter) sorted last:

    INNER JOIN reports ON reports.id = planned_disbursements.report_id
    ...
    ORDER BY
      ...
      reports.financial_year DESC NULLS LAST,
      reports.financial_quarter DESC NULLS LAST,
      reports.created_at DESC

For level A/B activities the query is simpler. Forecasts at this level
are not linked to reports, and we pick the latest version by sorting on
`planned_disbursement_type`:

    SELECT DISTINCT ON (
      planned_disbursements.financial_year,
      planned_disbursements.financial_quarter
    )
    planned_disbursements.*
    FROM planned_disbursements
    WHERE
      planned_disbursements.parent_activity_id = ?
    ORDER BY
      planned_disbursements.financial_year ASC,
      planned_disbursements.financial_year ASC,
      planned_disbursements.planned_disbursement_type DESC

To make this work for multiple Activity IDs at once, we need to change
the `WHERE` clause to allow a set of IDs instead of a single value:

    SELECT ...
    FROM planned_disbursements
    WHERE
      planned_disbursements.parent_activity_id IN (?)

Because we're selecting across multiple activities, we need to include
`parent_activity_id` in the `DISTINCT ON` clause, so we select a single
version for each activity and quarter.

    SELECT DISTINCT ON (
      planned_disbursements.parent_activity_id,
      planned_disbursements.financial_year,
      planned_disbursements.financial_quarter
    )
    ...
    ORDER BY
      planned_disbursements.parent_activity_id ASC,
      planned_disbursements.financial_year ASC,
      planned_disbursements.financial_year ASC,
      ...

Finally we need to deal with the fact these activities may be a mix of
level A/B and level C/D. Some will belong to a Report and be sorted on
the Report's quarter, and others will not. That means we need to use a
`LEFT OUTER JOIN` to prevent level A/B forecasts being excluded from the
results, and we need to add `planned_disbursement_type` to the `ORDER
BY` clause as a fallback for when none of the `reports` columns are
present.

    LEFT OUTER JOIN reports ON reports.id = planned_disbursements.report_id
    ...
    ORDER BY
      ...
      reports.financial_year DESC NULLS LAST,
      reports.financial_quarter DESC NULLS LAST,
      reports.created_at DESC,
      planned_disbursements.planned_disbursement_type DESC

Putting that all together gives us this query that can select the latest
forecast versions across a set of activities:

    SELECT DISTINCT ON (
      planned_disbursements.parent_activity_id,
      planned_disbursements.financial_year,
      planned_disbursements.financial_quarter
    )
    planned_disbursements.*
    FROM planned_disbursements
    LEFT OUTER JOIN reports ON reports.id = planned_disbursements.report_id
    WHERE
      planned_disbursements.parent_activity_id IN (?)
    ORDER BY
      planned_disbursements.parent_activity_id ASC,
      planned_disbursements.financial_year ASC,
      planned_disbursements.financial_year ASC,
      reports.financial_year DESC NULLS LAST,
      reports.financial_quarter DESC NULLS LAST,
      reports.created_at DESC,
      planned_disbursements.planned_disbursement_type DESC