Use the PlannedDisbursement{History,Overview} services to access forecasts
Most of the application and test suite should not access the
PlannedDisbursement model directly. It is a low-level implementation of
a data structure that represents versions of forecasts over time, and
reading and writing it directly will probably give weird or incorrect
results.

Instead, the PlannedDisbursementHistory service should be used to create
and update forecasts, and PlannedDisbursementOverview should be used to
read them. These services implement a high-level API on top of the
model, and the model and associated database table should be regarded as
an internal implementation detail.

This also means that FactoryBot should not be used to generate
PlannedDisbursement objects, since records in that table need to meet a
certain set of constraints in order to make sense. The
PlannedDisbursementHistory class should be used to create forecast data,
and the ReportingCycle service can help with this in tests by driving
the process of opening reports, approving them, and creating a report in
the following quarter.

This commit updates all the tests that should not access
PlannedDisbursement directly, to use the higher level services instead.