Use `PlannedDisbursement.unscoped` for direct model access
We plan to add a default scope to PlannedDisbursement that raises an
exception, so that any attempt to use the model directly will give an
error and tell users how to read and write forecases using the
PlannedDisbursementHistory and PlannedDisbursementOverview classes.

Before we do that, we need to mark any pieces of code that should
directly access the model so that they would bypass this default scope.
We do this by calling `PlannedDisbursement.unscoped` to get direct
access to the table data.