Stop code accessing PlannedDisbursement directly
We've had a few issues with this model being used directly when
PlannedDisbursementHistory and PlannedDisbursementOverview should be
used instead. Directly accessing the underlying data for this versioned
data type may give surprising or incorrect results, or result in invalid
data being created.

To prevent direct access, we add a default scope to the model that
raises an exception, giving advice about what to do instead. Code that
genuinely needs direct access to this model can use
`PlannedDisbursement.unscoped` to bypass the default scope.