Infer the value of `sdgs_apply` when importing
By default `sdgs_apply` is `false`.

Set it to true if at least one Sustainable Development Goal is populated during import.

NB: We don’t have a way to remove attributes during import, which is why we can’t currently detect if the goals are entirely removed in order to set it to `false`.