Default organisation's role to delivery partner
The previous commit added a new `role` column, but
was not set to anything by default. In later commits,
this can also be set to a matched effort provider, but
for now this will default to `delivery_partner` (there is
only ever one `service_provider`, and this is set outside
of the application frontend).

I also noticed that a feature test we had wasn't actually
working (It wasn't setting the
`beis_organisation_reference`, so the organisation was
invalid), so I've fixed that too.