Remove `planned_disbursement` factory
This factory is no longer used in the test suite and can be removed.
Going further than that, it's very difficult to use FactoryBot to create
reliably valid `planned_disbursements` records because of the
implementation details of that model. The PlannedDisbursementHistory and
ReportingCycle services should always be used to create forecasts in
tests.