Add a trigger rebase action
As well as triggering a rebase by setting a PR to
automerge or commenting `/rebase`, this PR adds the
functionality to check over all PRs that are set to
automerge and rebase them once a PR has been
merged to `develop` or `master`. This is particularly
helpful when we have a big backlog of PRs (i.e. from
Dependabot), so we don't have to constantly manually
rebase every time a new PR is merged.