Replace delivery_partner boolean with a role enum
This replaces the `delivery_partner` boolean column
with a `role` enum. This also has the side-effect of
making the default organisation created by a factory
neither a delivery partner or a service owner, which
doesn't reflect the reality of the application.

I've got round this by making the having the
root`organisation` factory "private", by adding a double
underscore at the start, in a similar way to what we've
already done with the Activities. Now, when an organisation
is used, it is either a delivery partner or a service owner.