Activity policy re: creating transfers
We need a way to check permissions for records that don't exist yet,
for example when deciding whether to show in the interface a button
for creating a transfer or not.

Whether a user is allowed to create a transfer from a source activity
depends on things we know at the level of this activity, and therefore
fits within the activity policy.

This is not typical to the way pundit encourages its use, but we are
using a similar pattern to check the permission to add a child activity
to a known activity.

Co-authored-by: jcoglan <james.coglan@dxw.com>