TransactionUploadsController uses upload? check
Use our new `ReportPolicy#upload?` check. Also, added controller specs
that checks that the 'upload' button appears at appropriate times.

When viewing the transaction upload page associated with a report that
is currently not editable, a message is shown instead of the upload
form.

NB: It is highly unlikely users will see this message, as the upload
buttons will soon be removed from the Report pages too.