Define what editable means for a report
We use this check a lot, so we want to define it and reuse it,
to allow for maintainability in case the definition of `editable` changes.

NB: The states are strings, not symbols. This wasn't causing a problem for
the query, but becomes a necessary distinction for an instance method.

Co-authored-by: jcoglan <james.coglan@dxw.com>