Apply transfer policies
In previous commits, we defined the transfer policy and
the activity policy relating to creating transfers from a source activity.

We are now using these policies across the app.

We are also showing transfers to all users who can access the activity's financials.
(Previously the entire transfers section was restricted to BEIS users.)

Restrict whether we show the create button or not according to the policy.

We think it's ok to not expand the controller spec to cover all scenarios, since
we now have feature specs to cover them.

Co-authored-by: jcoglan <james.coglan@dxw.com>