Report pages show the state
It's not obvious why the upload buttons show for some reports and not
others, so including the report's state in the header is a good first
step.

We could do more, but we'll shortly be overhauling the page layouts
throughout RODA and this will get improvements as part of that work.