Remove misleading 'administrator?' from policies
Originally, we thought users were going to have different access levels
to RODA, so we included a 'role' attribute in the User model. In
reality, *all* users are administrators, restricted by the organisation
they belong to:

- BEIS users can access all organisations
- Delivery partners can access their own organisation's data

Removing the `user.administrator?` calls, and still having passing tests
proves that this can be removed entirely.