Sum up forecasts by quarter
Reporting at level B (AKA programme) needs to add up all the forecasts
for all the child activities of the programme, as well as those at the
programme level itself, and then regroup them by their financial quarter.

The quarters are presented sorted, with the most recent at the top.

All the quarters that have forecasts reported are included. If there are quarters
without any forecasts reported, they won't be included. (No "zero reporting" even
if that means some quarters are "skipped".)

As the data isn't really a "planned disbursement" with a standalone history,
we are aggregating it in a temporary object that has the same interface as
a planned disbursement.

Co-authored-by: CristinaRO <cristina@dxw.com>